requirements.txt
# For TensorFlow on M1 Mac
tensorflow-macos>=2.12.0
tensorflow-metal>=0.8.0  

torch>=2.0.0
torchvision>=0.15.0

numba>=0.56.0
numpy>=1.24.0
librosa>=0.10.0